represent route array lengths as size_t
This fixes one -Wsign-conversion warning, though introduces a -Wconversion
warning. We should be able to remove the latter as we continue cleaning up the
use of int for sizes.

+++ lib/ortho/structures.h
@@ -14,6 +14,7 @@
 #include "types.h"
 #include "cgraph.h"
 #include <ortho/rawgraph.h>
+#include <stddef.h>
 
 typedef struct {
     double p1, p2;
@@ -45,7 +46,7 @@ typedef struct segment {
 } segment;
 
 typedef struct {
-  int n;
+  size_t n;
   segment* segs;
 } route;
